requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
pyparsing==3.0.9
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
